Simplified preparation for Hello, World!
In this rewrite of the tutorial preparation, we tell the user ahead of time what he/she will be doing. Also, much of the Read Ahead of Time reading goes to Suggested Additional Reading at the end. And also, the node.js and kubectrl downloads are no longer optional, as they are what the tutorial uses. A user who is sophisticated enough to dispense with them will know that he/she may do so without us saying so.